Move subshell syntax error test case to new file
